refactor: move generated api to shared infrastructure
