+++ src/features/reminder/presentation/hooks/useReminderData.ts
@@ -1,7 +1,7 @@
 import { useMemo } from 'react';
 import { mapServerDataToReminderSections } from '@src/features/reminder/domain/reminderMapper';
-import { useSectionsControllerFindAll } from '@src/features/auth/infrastructure/api/sections-섹션/sections-섹션';
-import { useRemindersControllerFindAll } from '@src/features/auth/infrastructure/api/reminders-리마인더/reminders-리마인더';
+import { useSectionsControllerFindAll } from '@src/shared/infrastructure/api/sections-섹션/sections-섹션';
+import { useRemindersControllerFindAll } from '@src/shared/infrastructure/api/reminders-리마인더/reminders-리마인더';
 
 export function useReminderData() {
   const { data: sectionsData, isPending: isPendingSections } = useSectionsControllerFindAll();
