+++ src/shared/infrastructure/api/model/sectionEntity.ts
