+++ src/shared/infrastructure/api/sections-섹션/sections-섹션.ts
@@ -30,7 +30,7 @@ import type {
   UpdateSectionDto
 } from '../model';
 
-import { customInstance } from '../../../../../shared/infrastructure/apiClient';
+import { customInstance } from '../../apiClient';
 
 
 type SecondParameter<T extends (...args: never) => unknown> = Parameters<T>[1];
