+++ src/shared/infrastructure/api/사용자-users/사용자-users.ts
@@ -24,7 +24,7 @@ import type {
   UserEntity
 } from '../model';
 
-import { customInstance } from '../../../../../shared/infrastructure/apiClient';
+import { customInstance } from '../../apiClient';
 
 
 type SecondParameter<T extends (...args: never) => unknown> = Parameters<T>[1];
