+++ src/shared/infrastructure/api/시스템-system/시스템-system.ts
@@ -20,7 +20,7 @@ import type {
   UseQueryResult
 } from '@tanstack/react-query';
 
-import { customInstance } from '../../../../../shared/infrastructure/apiClient';
+import { customInstance } from '../../apiClient';
 
 
 type SecondParameter<T extends (...args: never) => unknown> = Parameters<T>[1];
